requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-database-view==0.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-font-awesome==4.6.1
django-icons==4.0.0
django-import-export==2.5.0
et-xmlfile==1.1.0
fontawesome-free==5.15.4
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.8
psycopg2==2.9.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
